move common parts of sim.h to new file simcore.h
