+++ mosteksim/srcsim/Makefile.cygwin
@@ -88,83 +88,89 @@ allclean: clean
 
 # AUTOMATICALLY GENERATED DO NOT EDIT
 
-config.o: config.c sim.h $(CORE_DIR)/simglb.h $(CORE_DIR)/log.h
+config.o: config.c sim.h $(CORE_DIR)/simcore.h $(CORE_DIR)/simglb.h \
+ $(CORE_DIR)/log.h
 	$(CC) $(CFLAGS) -c -o $@ config.c
 
-iosim.o: iosim.c sim.h $(CORE_DIR)/simglb.h $(IO_DIR)/simbdos.h \
- $(IO_DIR)/mostek-cpu.h $(IO_DIR)/mostek-fdc.h \
- $(CORE_DIR)/log.h
+iosim.o: iosim.c sim.h $(CORE_DIR)/simcore.h $(CORE_DIR)/simglb.h \
+ $(IO_DIR)/simbdos.h $(IO_DIR)/mostek-cpu.h \
+ $(IO_DIR)/mostek-fdc.h $(CORE_DIR)/log.h
 	$(CC) $(CFLAGS) -c -o $@ iosim.c
 
-memory.o: memory.c sim.h $(CORE_DIR)/simglb.h memory.h
+memory.o: memory.c sim.h $(CORE_DIR)/simcore.h $(CORE_DIR)/simglb.h \
+ memory.h
 	$(CC) $(CFLAGS) -c -o $@ memory.c
 
-mostek-cpu.o: $(IO_DIR)/mostek-cpu.c sim.h $(CORE_DIR)/simglb.h \
- $(IO_DIR)/unix_terminal.h
+mostek-cpu.o: $(IO_DIR)/mostek-cpu.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h $(IO_DIR)/unix_terminal.h
 	$(CC) $(CFLAGS) -c -o $@ $(IO_DIR)/mostek-cpu.c
 
-mostek-fdc.o: $(IO_DIR)/mostek-fdc.c sim.h $(CORE_DIR)/simglb.h \
- $(CORE_DIR)/log.h
+mostek-fdc.o: $(IO_DIR)/mostek-fdc.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h $(CORE_DIR)/log.h
 	$(CC) $(CFLAGS) -c -o $@ $(IO_DIR)/mostek-fdc.c
 
-sim0.o: $(CORE_DIR)/sim0.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim0.o: $(CORE_DIR)/sim0.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim0.c
 
-sim1.o: $(CORE_DIR)/sim1.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim1.o: $(CORE_DIR)/sim1.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim1.c
 
-sim1a.o: $(CORE_DIR)/sim1a.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim1a.o: $(CORE_DIR)/sim1a.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim1a.c
 
-sim2.o: $(CORE_DIR)/sim2.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim2.o: $(CORE_DIR)/sim2.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim2.c
 
-sim3.o: $(CORE_DIR)/sim3.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim3.o: $(CORE_DIR)/sim3.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim3.c
 
-sim4.o: $(CORE_DIR)/sim4.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim4.o: $(CORE_DIR)/sim4.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim4.c
 
-sim5.o: $(CORE_DIR)/sim5.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim5.o: $(CORE_DIR)/sim5.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim5.c
 
-sim6.o: $(CORE_DIR)/sim6.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim6.o: $(CORE_DIR)/sim6.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim6.c
 
-sim7.o: $(CORE_DIR)/sim7.c sim.h $(CORE_DIR)/simglb.h config.h \
- memory.h
+sim7.o: $(CORE_DIR)/sim7.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h config.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/sim7.c
 
-simbdos.o: $(IO_DIR)/simbdos.c sim.h $(CORE_DIR)/simglb.h \
- memory.h
+simbdos.o: $(IO_DIR)/simbdos.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(IO_DIR)/simbdos.c
 
-simctl.o: simctl.c sim.h $(CORE_DIR)/simglb.h \
+simctl.o: simctl.c sim.h $(CORE_DIR)/simcore.h $(CORE_DIR)/simglb.h \
  $(IO_DIR)/mostek-fdc.h $(IO_DIR)/unix_terminal.h
 	$(CC) $(CFLAGS) -c -o $@ simctl.c
 
-simdis.o: $(CORE_DIR)/simdis.c sim.h $(CORE_DIR)/simglb.h memory.h
+simdis.o: $(CORE_DIR)/simdis.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/simdis.c
 
-simfun.o: $(CORE_DIR)/simfun.c sim.h $(CORE_DIR)/simglb.h memory.h \
- $(CORE_DIR)/log.h
+simfun.o: $(CORE_DIR)/simfun.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h memory.h $(CORE_DIR)/log.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/simfun.c
 
-simglb.o: $(CORE_DIR)/simglb.c sim.h $(CORE_DIR)/simglb.h
+simglb.o: $(CORE_DIR)/simglb.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/simglb.c
 
-simice.o: $(CORE_DIR)/simice.c sim.h $(CORE_DIR)/simglb.h memory.h
+simice.o: $(CORE_DIR)/simice.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h memory.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/simice.c
 
-simint.o: $(CORE_DIR)/simint.c sim.h $(CORE_DIR)/simglb.h
+simint.o: $(CORE_DIR)/simint.c sim.h $(CORE_DIR)/simcore.h \
+ $(CORE_DIR)/simglb.h
 	$(CC) $(CFLAGS) -c -o $@ $(CORE_DIR)/simint.c
 
 unix_terminal.o: $(IO_DIR)/unix_terminal.c
